// controls/richtexteditor/blazor-script/editor-scripts/common/interface.ts
export interface IDropDownItemModel extends DropDownItemModel {
    cssClass?: string
    command?: string
    subCommand?: string
    value?: string
    text?: string
}

/**
 * @hidden
 * @deprecated
 */
export interface IToolbarItemModel extends ItemModel {
    command?: string
    subCommand?: string
}
// ...
export interface IShowPopupArgs {
    args?: MouseEvent | TouchEvent | KeyboardEvent
    type?: string
    isNotify: boolean
    elements?: Element | Element[]
}
// ...
export interface DialogCloseEventArgs {
    /**
     * Identifies if the current action can be canceled.
     */
    cancel: boolean
    /**
     * Returns the root container element of the dialog being closed.
     */
    container: HTMLElement
    /**
     * Provides reference to the dialog element being closed.
     */
    element: Element
    /**
     * Returns the original event arguments, if any.
     */
    event: Event
    /**
     * Determines if the dialog close event is triggered by user interaction.
     */
    isInteracted: boolean
    /**
     * DEPRECATED-Determines whether the event is triggered by interaction.
     */
    isInteraction: boolean
    /**
     * Specifies the event name, if available.
     */
    /* eslint-disable */
    name?: String
    /* eslint-enable */
    /**
     * Determines if action can be prevented; target details.
     */
    /* eslint-disable */
    target: HTMLElement | String
    /* eslint-enable */
}